add function to generate password to specific folder
